server/requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
pwdlib[argon2]>=0.2.0

# TTS and audio processing
kokoro>=0.9.4
soundfile>=0.12.1
pydub>=0.25.1
num2words>=0.5.13

# Document processing
pymupdf4llm>=0.0.5
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0

# Testing (optional)
pytest>=9.0.0
pytest-asyncio>=1.3.0
